# paymentsService/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: API для управления балансом на счете пользователя и платежами
  title: Payment Service API
  version: 1.0.0
servers:
- url: /
paths:
  /users/{userId}:
    get:
      parameters:
      - description: Идентификатор пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Balance"
                type: array
          description: Успешный ответ
        "404":
          description: Пользователь не найден
        "5XX":
          description: Внутренняя ошибка сервера
        default:
          description: Неизвестная ошибка при выяснении баланса
      summary: Получить баланс пользователя
      x-accepts:
      - application/json
  /users/{userId}/makePayment:
    post:
      parameters:
      - explode: true
        in: query
        name: sum
        required: true
        schema:
          example: 101.15
          format: float
          type: number
        style: form
      - description: Идентификатор пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
          description: Успешный ответ
        "404":
          description: Пользователь не найден
        "5XX":
          description: Внутренняя ошибка сервера
        default:
          description: Неизвестная ошибка при выполнении платежа
      summary: Совершить платеж (совершить списание денег со счёта)
      x-accepts:
      - application/json
components:
  schemas:
    PaymentResponse:
      example:
        restBalance: 0
        message: OK
      properties:
        restBalance:
          description: Остаточный баланс
          format: float
          type: integer
        message:
          description: Описание статуса ответа
          example: OK
          type: string
      type: object
    Balance:
      example:
        balance: 101.15
      properties:
        balance:
          example: 101.15
          format: float
          type: number
      type: object